fixed time table crash
